refactor: executeWithLatch 확장함수를 활용하도록 리팩토링
