+++ app/src/main/java/woowacourse/shopping/data/repository/real/RealProductRepositoryImpl.kt
@@ -6,35 +6,22 @@ import woowacourse.shopping.domain.model.Product
 import woowacourse.shopping.domain.repository.ProductRepository
 import woowacourse.shopping.utils.DtoMapper.toProduct
 import woowacourse.shopping.utils.DtoMapper.toProducts
-import woowacourse.shopping.utils.exception.LatchUtils.awaitOrThrow
+import woowacourse.shopping.utils.exception.LatchUtils.executeWithLatch
 import woowacourse.shopping.utils.exception.NoSuchDataException
-import java.util.concurrent.CountDownLatch
-import kotlin.concurrent.thread
 
 class RealProductRepositoryImpl(
     private val productDataSource: ProductDataSource = ProductDataSourceImpl(),
 ) : ProductRepository {
     override fun loadPagingProducts(offset: Int): List<Product> {
-        val latch = CountDownLatch(1)
         var products: List<Product>? = null
-        var exception: Exception? = null
-
-        thread {
-            try {
-                val page = offset / PRODUCT_LOAD_PAGING_SIZE
-                val response =
-                    productDataSource.loadProducts(page, PRODUCT_LOAD_PAGING_SIZE).execute()
-                if (response.isSuccessful && response.body() != null) {
-                    products = response.body()?.toProducts()
-                }
-            } catch (e: Exception) {
-                exception = e
-            } finally {
-                latch.countDown()
+        executeWithLatch {
+            val page = offset / PRODUCT_LOAD_PAGING_SIZE
+            val response =
+                productDataSource.loadProducts(page, PRODUCT_LOAD_PAGING_SIZE).execute()
+            if (response.isSuccessful && response.body() != null) {
+                products = response.body()?.toProducts()
             }
         }
-
-        latch.awaitOrThrow(exception)
         return products ?: throw NoSuchDataException()
     }
 
@@ -42,51 +29,29 @@ class RealProductRepositoryImpl(
         size: Int,
         category: String,
     ): List<Product> {
-        val latch = CountDownLatch(1)
         var products: List<Product>? = null
-        var exception: Exception? = null
-
-        thread {
-            try {
-                val response =
-                    productDataSource.loadCategoryProducts(
-                        page = DEFAULT_PAGE,
-                        size = RECOMMEND_PRODUCT_LOAD_PAGING_SIZE,
-                        category = category,
-                    ).execute()
-                if (response.isSuccessful && response.body() != null) {
-                    products = response.body()?.toProducts()
-                }
-            } catch (e: Exception) {
-                exception = e
-            } finally {
-                latch.countDown()
+        executeWithLatch {
+            val response =
+                productDataSource.loadCategoryProducts(
+                    page = DEFAULT_PAGE,
+                    size = RECOMMEND_PRODUCT_LOAD_PAGING_SIZE,
+                    category = category,
+                ).execute()
+            if (response.isSuccessful && response.body() != null) {
+                products = response.body()?.toProducts()
             }
         }
-        latch.awaitOrThrow(exception)
-
         return products ?: throw NoSuchDataException()
     }
 
     override fun getProduct(productId: Long): Product {
-        val latch = CountDownLatch(1)
         var product: Product? = null
-        var exception: Exception? = null
-
-        thread {
-            try {
-                val response = productDataSource.loadProduct(productId.toInt()).execute()
-                if (response.isSuccessful && response.body() != null) {
-                    product = response.body()?.toProduct()
-                }
-            } catch (e: Exception) {
-                exception = e
-            } finally {
-                latch.countDown()
+        executeWithLatch {
+            val response = productDataSource.loadProduct(productId.toInt()).execute()
+            if (response.isSuccessful && response.body() != null) {
+                product = response.body()?.toProduct()
             }
         }
-
-        latch.awaitOrThrow(exception)
         return product ?: throw NoSuchDataException()
     }
